+++ src/Validators/ValidatorErrorFactory.php
@@ -18,9 +18,11 @@
 
 namespace CloudCreativity\JsonApi\Validators;
 
+use CloudCreativity\JsonApi\Contracts\Document\MutableErrorInterface;
 use CloudCreativity\JsonApi\Contracts\Object\ResourceIdentifierInterface;
 use CloudCreativity\JsonApi\Contracts\Repositories\ErrorRepositoryInterface;
 use CloudCreativity\JsonApi\Contracts\Validators\ValidatorErrorFactoryInterface;
+use CloudCreativity\JsonApi\Exceptions\MutableErrorCollection;
 use CloudCreativity\JsonApi\Utils\Pointer as P;
 
 /**
@@ -206,13 +208,26 @@ class ValidatorErrorFactory implements ValidatorErrorFactoryInterface
     /**
      * @inheritdoc
      */
-    public function relationshipNotAcceptable(ResourceIdentifierInterface $identifier, $relationshipKey = null)
-    {
-        return $this->repository->errorWithPointer(
+    public function relationshipNotAcceptable(
+        ResourceIdentifierInterface $identifier,
+        $relationshipKey = null,
+        $error = null
+    ) {
+        $base = $this->repository->errorWithPointer(
             self::RELATIONSHIP_NOT_ACCEPTABLE,
             $relationshipKey ? P::relationship($relationshipKey) : P::data(),
             ['type' => $identifier->getType(), 'id' => $identifier->getId()]
         );
+
+        $errors = new MutableErrorCollection();
+
+        /** @var MutableErrorInterface $err */
+        foreach (MutableErrorCollection::cast($error ?: $base) as $err) {
+            $add = clone $base;
+            $errors->add($add->merge($err));
+        }
+
+        return $errors;
     }
 
 }